requirements.txt
streamlit>=1.44
pandas>=2.2
openpyxl>=3.1
requests>=2.32
feedparser>=6.0
beautifulsoup4>=4.12
python-dateutil>=2.9
PyYAML>=6.0
